+++ src/Styleguide/Components/__stories__/ArtistBio.story.tsx
@@ -2,28 +2,7 @@ import React from "react"
 import { storiesOf } from "storybook/storiesOf"
 import { Section } from "Styleguide/Utils/Section"
 import { ArtistBio } from "../ArtistBio"
-
-export const bio = `Donald Judd, widely regarded as one of the most significant American artists of
-the post-war period, is perhaps best-known for the large-scale outdoor
-installations and long, spacious interiors he designed in Marfa, Texas. His
-oeuvre has come to define what has been referred to as Minimalist art—a label
-the artist strongly objected to. His sculptures and installations, constructed
-out of industrial materials such as Plexiglas, concrete, and steel and arranged
-in precise geometric shapes, were intended to emphasize the purity of the
-objects themselves rather than any symbolic meaning they might have—“the simple
-expression of complex thought,” said Judd. His particular interest in
-architecture led him to design both the sculptures and the spaces in which they
-would be contained. Donald Judd, widely regarded as one of the most significant
-American artists of the post-war period, is perhaps best-known for the
-large-scale outdoor installations and long, spacious interiors he designed in
-Marfa, Texas. His oeuvre has come to define what has been referred to as
-Minimalist art—a label the artist strongly objected to. His sculptures and
-installations, constructed out of industrial materials such as Plexiglas,
-concrete, and steel and arranged in precise geometric shapes, were intended to
-emphasize the purity of the objects themselves rather than any symbolic meaning
-they might have—“the simple expression of complex thought,” said Judd. His
-particular interest in architecture led him to design both the sculptures and
-the spaces in which they would be contained.`
+import { bio } from "Styleguide/Pages/Fixtures/ArtistBio"
 
 storiesOf("Styleguide/Components", module).add("ArtistBio", () => {
   return (
