+++ src/Styleguide/Components/__stories__/MarketInsight.story.tsx
@@ -2,21 +2,7 @@ import React from "react"
 import { storiesOf } from "storybook/storiesOf"
 import { Section } from "Styleguide/Utils/Section"
 import { MarketInsights } from "../MarketInsights"
-
-export const insights = [
-  {
-    primaryLabel: "High auction record",
-    secondaryLabel: "$105m sale, Sotheby’s 2013",
-  },
-  {
-    primaryLabel: "Blue Chip",
-    secondaryLabel: "Represented by internationally recognized galleries.",
-  },
-  {
-    primaryLabel: "Collected by major museums",
-    secondaryLabel: "Tate, Museum of Modern Art (MoMA)",
-  },
-]
+import { insights } from "Styleguide/Pages/Fixtures/MarketInsights"
 
 storiesOf("Styleguide/Components", module).add("MarketInsights", () => {
   return (
