+++ src/Styleguide/Components/__stories__/SelectedExhibitions.story.tsx
@@ -5,24 +5,7 @@ import {
   SelectedExhibitionsContainer,
 } from "../SelectedExhibitions"
 import { Section } from "Styleguide/Utils/Section"
-
-export const exhibitions = [
-  {
-    year: "2018",
-    show: "Adman: Warhol Before Pop",
-    gallery: "Andy Warhol Museum",
-  },
-  {
-    year: "2018",
-    show: "Brancusi: Pioneer of American Minimalism",
-    gallery: "Paul Kasmin Gallery",
-  },
-  {
-    year: "2017",
-    show: "Sculpture on the Move 1946–2016",
-    gallery: "Kunstmuseum Basel",
-  },
-]
+import { exhibitions } from "Styleguide/Pages/Fixtures/SelectedExhibitions"
 
 storiesOf("Styleguide/Components", module).add("SelectedExhibitions", () => {
   return (
