+++ src/Styleguide/Pages/Artist/index.js
@@ -1,6 +1,5 @@
 import React from "react"
 import { Tabs, Tab } from "Styleguide/Components/Tabs"
-import { Box } from "Styleguide/Elements/Box"
 import { Col, Grid, Row } from "Styleguide/Elements/Grid"
 import { Separator } from "Styleguide/Elements/Separator"
 import { Spacer } from "Styleguide/Elements/Spacer"
