+++ src/Styleguide/Pages/Artwork/__stories__/Sidebar/Artists.story.tsx
@@ -2,41 +2,11 @@ import React from "react"
 import { storiesOf } from "storybook/storiesOf"
 import { Section } from "Styleguide/Utils/Section"
 import { Artists } from "../../Sidebar/Artists"
-
-const SingleFollowedArtist = [
-  {
-    __id: "francesca-dimattio",
-    id: "francesca-dimattio",
-    name: "Francesca DiMattio",
-    is_followed: true,
-    href: "https://www.artsy.net/artist/francesca-dimattio",
-  },
-]
-
-const SingleNonFollowedArtist = [
-  {
-    __id: "francesca-dimattio",
-    id: "francesca-dimattio",
-    name: "Francesca DiMattio",
-    is_followed: false,
-  },
-]
-
-const MultipleArtists = [
-  {
-    __id: "francesca-dimattio",
-    id: "francesca-dimattio",
-    name: "Francesca DiMattio",
-    is_followed: false,
-  },
-  {
-    __id: "sol-lewitt-piramidi-c",
-    id: "sol-lewitt-piramidi-c",
-    name: "Sol LeWitt",
-    href: "https://www.artsy.net/artwork/sol-lewitt-piramidi-c",
-    is_followed: false,
-  },
-]
+import {
+  SingleFollowedArtist,
+  SingleNonFollowedArtist,
+  MultipleArtists,
+} from "Styleguide/Pages/Fixtures/Artwork/Sidebar/Artists"
 
 storiesOf("Styleguide/Artwork/Sidebar", module).add("Artists", () => {
   return (
