+++ src/Styleguide/Pages/Artwork/__stories__/Sidebar/SizeInfo.story.tsx
@@ -2,20 +2,11 @@ import React from "react"
 import { storiesOf } from "storybook/storiesOf"
 import { Section } from "Styleguide/Utils/Section"
 import { SizeInfo } from "../../Sidebar/SizeInfo"
-
-const ArtworkWithSizeAndEditionOf = {
-  dimensions: { in: "10 1/4 × 8 7/8 in", cm: "26 × 22.5 cm" },
-  edition_of: "Edition of 20",
-}
-
-const ArtworkWithSizeOnly = {
-  dimensions: { in: "10 1/4 × 8 7/8 in", cm: "26 × 22.5 cm" },
-}
-
-const ArtworkWithEditionOfOnly = {
-  dimensions: { in: null, cm: null },
-  edition_of: "Edition of 20",
-}
+import {
+  ArtworkWithSizeAndEditionOf,
+  ArtworkWithSizeOnly,
+  ArtworkWithEditionOfOnly,
+} from "Styleguide/Pages/Fixtures/Artwork/Sidebar/SizeInfo"
 
 storiesOf("Styleguide/Artwork/Sidebar", module).add("SizeInfo", () => {
   return (
