+++ src/Styleguide/Pages/Fixtures/ArtistBio.ts
@@ -1,0 +1,21 @@
+export const bio = `Donald Judd, widely regarded as one of the most significant American artists of
+the post-war period, is perhaps best-known for the large-scale outdoor
+installations and long, spacious interiors he designed in Marfa, Texas. His
+oeuvre has come to define what has been referred to as Minimalist art—a label
+the artist strongly objected to. His sculptures and installations, constructed
+out of industrial materials such as Plexiglas, concrete, and steel and arranged
+in precise geometric shapes, were intended to emphasize the purity of the
+objects themselves rather than any symbolic meaning they might have—“the simple
+expression of complex thought,” said Judd. His particular interest in
+architecture led him to design both the sculptures and the spaces in which they
+would be contained. Donald Judd, widely regarded as one of the most significant
+American artists of the post-war period, is perhaps best-known for the
+large-scale outdoor installations and long, spacious interiors he designed in
+Marfa, Texas. His oeuvre has come to define what has been referred to as
+Minimalist art—a label the artist strongly objected to. His sculptures and
+installations, constructed out of industrial materials such as Plexiglas,
+concrete, and steel and arranged in precise geometric shapes, were intended to
+emphasize the purity of the objects themselves rather than any symbolic meaning
+they might have—“the simple expression of complex thought,” said Judd. His
+particular interest in architecture led him to design both the sculptures and
+the spaces in which they would be contained.`
