+++ src/Styleguide/Pages/Fixtures/Artwork/Sidebar/Artists.ts
@@ -1,0 +1,34 @@
+export const SingleFollowedArtist = [
+  {
+    __id: "francesca-dimattio",
+    id: "francesca-dimattio",
+    name: "Francesca DiMattio",
+    is_followed: true,
+    href: "https://www.artsy.net/artist/francesca-dimattio",
+  },
+]
+
+export const SingleNonFollowedArtist = [
+  {
+    __id: "francesca-dimattio",
+    id: "francesca-dimattio",
+    name: "Francesca DiMattio",
+    is_followed: false,
+  },
+]
+
+export const MultipleArtists = [
+  {
+    __id: "francesca-dimattio",
+    id: "francesca-dimattio",
+    name: "Francesca DiMattio",
+    is_followed: false,
+  },
+  {
+    __id: "sol-lewitt-piramidi-c",
+    id: "sol-lewitt-piramidi-c",
+    name: "Sol LeWitt",
+    href: "https://www.artsy.net/artwork/sol-lewitt-piramidi-c",
+    is_followed: false,
+  },
+]
