+++ src/Styleguide/Pages/Fixtures/Artwork/Sidebar/Classification.ts
@@ -1,0 +1,5 @@
+export const ArtworkWithClassification = {
+  attribution_class: { short_description: "This is a unique work" },
+}
+
+export const ArtworkWithoutClassification = { attribution_class: null }
