+++ src/Styleguide/Pages/Fixtures/Artwork/Sidebar/SizeInfo.ts
@@ -1,0 +1,13 @@
+export const ArtworkWithSizeAndEditionOf = {
+  dimensions: { in: "10 1/4 × 8 7/8 in", cm: "26 × 22.5 cm" },
+  edition_of: "Edition of 20",
+}
+
+export const ArtworkWithSizeOnly = {
+  dimensions: { in: "10 1/4 × 8 7/8 in", cm: "26 × 22.5 cm" },
+}
+
+export const ArtworkWithEditionOfOnly = {
+  dimensions: { in: null, cm: null },
+  edition_of: "Edition of 20",
+}
