+++ src/Styleguide/Pages/Fixtures/Artwork/Sidebar/TitleInfo.ts
@@ -1,0 +1,21 @@
+export const ArtworkWithTitleDateAndMedium = {
+  title: "The Fox and the Hound",
+  date: "2018",
+  medium: "Oil on canvas",
+}
+
+export const ArtworkWithTitleOnly = {
+  title: "The Fox and the Hound",
+}
+
+export const ArtworkWithTitlAndDate = {
+  title: "The Fox and the Hound",
+  date: " 2013 - 2012 ",
+}
+
+export const ArtworkWithTitleAndMedium = {
+  title: "The Fox and the Hound",
+  date: null,
+  medium:
+    "Hand selected materials to paint this piece were sourced from the old villages on magic lands.",
+}
