+++ src/Styleguide/Pages/Fixtures/MarketInsights.ts
@@ -1,0 +1,14 @@
+export const insights = [
+  {
+    primaryLabel: "High auction record",
+    secondaryLabel: "$105m sale, Sotheby’s 2013",
+  },
+  {
+    primaryLabel: "Blue Chip",
+    secondaryLabel: "Represented by internationally recognized galleries.",
+  },
+  {
+    primaryLabel: "Collected by major museums",
+    secondaryLabel: "Tate, Museum of Modern Art (MoMA)",
+  },
+]
