+++ src/Styleguide/Pages/Fixtures/SelectedExhibitions.ts
@@ -1,0 +1,17 @@
+export const exhibitions = [
+  {
+    year: "2018",
+    show: "Adman: Warhol Before Pop",
+    gallery: "Andy Warhol Museum",
+  },
+  {
+    year: "2018",
+    show: "Brancusi: Pioneer of American Minimalism",
+    gallery: "Paul Kasmin Gallery",
+  },
+  {
+    year: "2017",
+    show: "Sculpture on the Move 1946–2016",
+    gallery: "Kunstmuseum Basel",
+  },
+]
